requirements.txt
tensorflow==2.1.0
bert-for-tf2
tensorflow-datasets
tqdm
numpy

